vpu: v2: fix cache missing in ctx switching
This bug is caused from the code optimization. That is why it
could happen in a platform while the other is free of it with
the same code.

Also we need to enclose the context switching, in some scope,
we can't hold a lock since a sleepable lock has been invoked.

JIRA:
VSSDK-29432

Change-Id: Ia1201ef5f219709207d5b20729b2776608708315
Signed-off-by: Hsia-Jun(Randy) Li <randy.li@synaptics.com>
Reviewed-on: https://sc-debu-git.synaptics.com/gerrit/219928
Reviewed-by: Randy Li <Randy.Li@synaptics.com>
Reviewed-by: Yanyan Peng <Yanyan.Peng@synaptics.com>