[DHD]: Patch for removing addon static interface
Per Yacto customer's below request, remove addon static interface in this patch.
------------------------------------------------------------------
Wondering why the astra-media yocto build is showing two wlan devices when only one wireless phy is present, and if it has no purpose then how best to remove it.
the wlan device keeps periodically scanning while wlan is connected, leading to a delay in traffic for
100-200ms for wlan which is not acceptable for my use case.
If wlan is for some advanced purpose such as mesh or backhaul, would prefer to remove it as I don't have that need.

Change-Id: Ib77e53a36ed724e219e909b8db88f8a34cb019eb
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/223608
Reviewed-by: Chunguang Wang <Chunguang.Wang@synaptics.com>
Reviewed-by: Rajashekar Mathivaanan <Rajashekar.Mathivaanan@synaptics.com>
Reviewed-by: Hong Song <Hong.Song@synaptics.com>
Reviewed-by: Jisheng Zhang <Jisheng.Zhang@synaptics.com>